+++ test/unit/index.js
@@ -62,6 +62,7 @@ require('./spec/signaling/v2/peerconnectionmanager');
 require('./spec/signaling/v2/localtrackpublication');
 require('./spec/signaling/v2/remotetrackpublication');
 require('./spec/signaling/v2/trackprioritysignaling');
+require('./spec/signaling/v2/transcriptionsignaling');
 require('./spec/signaling/v2/twilioconnectiontransport');
 
 require('./spec/util');