+++ test/unit/spec/signaling/v2/twilioconnectiontransport.js
@@ -19,6 +19,7 @@ describe('TwilioConnectionTransport', () => {
     [true, false], // iceServers
     [true, false], // networkQuality
     [true, false], // dominantSpeaker
+    [true, false], // receiveTranscriptions
     [true, false], // automaticSubscription
     [true, false], // trackPriority
     [true, false], // trackSwitchOff
@@ -89,6 +90,7 @@ describe('TwilioConnectionTransport', () => {
     iceServers,
     networkQuality,
     dominantSpeaker,
+    receiveTranscriptions,
     automaticSubscription,
     trackPriority,
     trackSwitchOff,
@@ -101,6 +103,7 @@ describe('TwilioConnectionTransport', () => {
       .iceServers ${iceServers ? '' : 'not '}provided
       .networkQuality flag ${networkQuality ? 'enabled' : 'disabled'},
       .dominantSpeaker flag ${dominantSpeaker ? 'enabled' : 'disabled'},
+      .receiveTranscriptions flag ${receiveTranscriptions ? 'enabled' : 'disabled'},
       .automaticSubscription flag ${automaticSubscription ? 'enabled' : 'disabled'},
       .trackPriority flag ${trackPriority ? 'enabled' : 'disabled'},
       .trackSwitchOff flag ${trackSwitchOff ? 'enabled' : 'disabled'},
@@ -117,6 +120,7 @@ describe('TwilioConnectionTransport', () => {
           automaticSubscription,
           networkQuality,
           dominantSpeaker,
+          receiveTranscriptions,
           trackPriority,
           trackSwitchOff,
           notifyWarnings
@@ -146,6 +150,7 @@ describe('TwilioConnectionTransport', () => {
         it(`should ${action} a Connect RSP message that
           ${networkQuality ? 'contains' : 'does not contain'} the "network_quality" payload,
           ${dominantSpeaker ? 'contains' : 'does not contain'} the "active_speaker" payload,
+          ${receiveTranscriptions ? 'contains' : 'does not contain'} the "extension_transcriptions" payload,
           the "subscribe-${automaticSubscription ? 'all' : 'none'}" subscription rule,
           ${trackPriority ? 'contains' : 'does not contain'} the "track_priority" payload,
           ${trackSwitchOff ? 'contains' : 'does not contain'} the "track_switch_off" payload, and
@@ -168,6 +173,12 @@ describe('TwilioConnectionTransport', () => {
             assert.deepEqual(message.media_signaling.active_speaker.transports, [{ type: 'data-channel' }]);
           } else {
             assert(!('active_speaker' in message.media_signaling));
+          }
+
+          if (receiveTranscriptions) {
+            assert.deepEqual(message.media_signaling.extension_transcriptions.transports, [{ type: 'data-channel' }]);
+          } else {
+            assert(!('extension_transcriptions' in message.media_signaling));
           }
 
           if (trackPriority) {