+++ tsdef/twilio-video-tests.ts
@@ -382,6 +382,10 @@ async function initRoom() {
       if (publication.track.kind !== 'data') { trackUnsubscribed(publication.track); }
     });
   });
+  room.on('transcription', (event: Video.TranscriptionEvent) => {
+    // eslint-disable-next-line no-console
+    console.log(`${event.participant}: ${event.transcription}`);
+  });
 }
 
 function unpublishTracks() {
@@ -432,6 +436,7 @@ function insertDomElement(element: HTMLMediaElement) {
 function useConnectionOptions() {
   const connectionOptions: Video.ConnectOptions = {
     dominantSpeaker: true,
+    receiveTranscriptions: true,
     networkQuality: { local: 1, remote: 1 },
     maxAudioBitrate: Number('13000'),
     preferredAudioCodecs: [{  codec: 'opus', dtx: false }],