+++ tsdef/types.d.ts
@@ -214,7 +214,7 @@ export interface ConnectOptions extends MediaImplementationOptions {
   automaticSubscription?: boolean;
   bandwidthProfile?: BandwidthProfileOptions;
   dominantSpeaker?: boolean;
-
+  receiveTranscriptions?: boolean;
   /**
    * @deprecated use enableDscp
    */
@@ -318,4 +318,17 @@ export class StatsReport {
 
 export interface CancelablePromise<T> extends Promise<T> {
   cancel: () => void;
+}
+
+export type TranscriptionEvent = {
+  /* eslint-disable camelcase */
+  language_code: string;
+  partial_results: boolean;
+  sequence_number: number;
+  /* eslint-enable camelcase */
+  participant: string;
+  timestamp: string;
+  track: string;
+  transcription: string;
+  type: 'extension_transcriptions';
 }